pyproject.toml
[project]
name = "llm-freeway"
version = "0.1.0"
description = ""
authors = [
    {name = "gecBurton",email = "g.e.c.burton@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "litellm (>=1.62.1,<2.0.0)",
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "sqlmodel (>=0.0.23,<0.0.24)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "google-auth (>=2.38.0,<3.0.0)",
    "boto3 (>=1.37.18,<2.0.0)",
    "python-keycloak (>=5.3.1,<6.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
trio = "^0.29.0"
ruff = "^0.9.9"
pytest-cov = "^6.0.0"
pytest-freezegun = "^0.4.2"
setuptools = "^76.1.0"

[tool.ruff.lint]
extend-select = ["I"]